% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLikes.R
\name{getLikes}
\alias{getLikes}
\title{Extract list of likes of the authenticated Facebook user}
\usage{
getLikes(user = "me", n = 500, token)
}
\arguments{
\item{user}{The name of the user. By default, "me".}

\item{n}{Maximum number of likes to return for each user.}

\item{token}{Either a temporary access token created at
\url{https://developers.facebook.com/tools/explorer} or the OAuth token 
created with \code{fbOAuth}.}
}
\description{
\code{getLikes} retrieves information about the authenticated use
only (not any friend). To retrieve the number of likes for a page, 
use \code{getUsers} with the page IDs.
}
\details{
This function requires the use of an OAuth token with the following
permission: user_likes
}
\examples{
\dontrun{
 token <- 'XXXXX'
 my_likes <- getLikes(user="me", token=token)
}

}
\seealso{
\code{\link{getFriends}}, \code{\link{fbOAuth}}
}
\author{
Pablo Barbera \email{pbarbera@usc.edu}
}
