% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPageInfo.R
\name{getPageInfo}
\alias{getPageInfo}
\title{Extract information about a public Facebook page}
\usage{
getPageInfo(page, token, fields = "name", api = NULL)
}
\arguments{
\item{page}{A page ID or page name.}

\item{token}{Either a temporary access token created at
\url{https://developers.facebook.com/tools/explorer} or the OAuth token 
created with \code{fbOAuth}.}

\item{fields}{Character vector of fields to extract, e.g. name, about, category, fan_count}
}
\description{
\code{getPage} retrieves information from a public Facebook page.
}
\details{
This function will only return information from public pages, not users
with public profiles.
}
\examples{
\dontrun{

}

}
\seealso{
\code{\link{getUsers}}, \code{\link{getPost}}, \code{\link{fbOAuth}}
}
\author{
Wouter van Atteveldt \email{wouter@vanatteveldt.com}
}
