% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getPermissions.R
\name{getPermissions}
\alias{getPermissions}
\title{Extract information about the permissions for the current user}
\usage{
getPermissions(token)
}
\arguments{
\item{token}{Either a temporary access token created at
\url{https://developers.facebook.com/tools/explorer} or the OAuth token
created with \code{fbOAuth}.}
}
\description{
\code{getPermissions} retrieves informations about the permissions the current user actually has.
}
\details{
\code{getPermissions} returns a character vector where each row is an action the authenticated user is granted to perform
through Facebook Graph API with the current token.
}
\examples{
\dontrun{
## See examples for fbOAuth to know how token was created.
## Getting information about current user's permissions.
	load("fb_oauth")
	me.permissions <- getPermissions(token=fb_oauth)
}
}
\author{
Gabriele Baldassarre
}
\seealso{
\code{\link{fbOAuth}}
}

