% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\title{Pproxt function to perform a generic query against FB Graph API}
\usage{
query(query, token)
}
\arguments{
\item{token}{Either a temporary access token created at
\url{https://developers.facebook.com/tools/explorer} or the OAuth token
created with \code{fbOAuth}.}
}
\description{
\code{query} performs a generic query against Facebook Graph API and returns a list reflecting the outputted json tree.
}
\details{
Write the query as you would do using the Graph Explroer at \url{https://developers.facebook.com/tools/explorer}
but don't include the API version number and the trailing slash.
Please note that this function has no automatic paging support.
}
\examples{
\dontrun{
## Copy and paste token created at FB Graph API Explorer
 token <- "XXXXXX"

	results.json <- query(query="me/posts/?limit=10", token=token)
}
}
\author{
Gabriele Baldassarre \email{gabriele@gabrielebaldassarre.com}
}
\seealso{
\code{\link{fbOAuth}}
}

