% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfacebook-package.R
\docType{package}
\name{Rfacebook-package}
\alias{Rfacebook-package}
\alias{Rfacebook}
\title{Access to Facebook API via R}
\description{
This package provides a series of functions that allow R users
to access Facebook's API to get information about users and posts,
and collect public status updates that mention specific keywords.
}
\seealso{
\code{\link{fbOAuth}}, \code{\link{getUsers}}, 
\code{\link{getPost}}, \code{\link{searchFacebook}},
\code{\link{updateStatus}}, \code{\link{getFriends}}, 
\code{\link{getNetwork}}, \code{\link{getPage}}
}
\author{
Pablo Barbera \email{pbarbera@usc.edu}
}
