% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/searchFacebook.R
\name{searchFacebook}
\alias{searchFacebook}
\title{Search public posts that mention a string}
\usage{
searchFacebook(string, token, n = 200, since = NULL, until = NULL,
  api = "2")
}
\arguments{
\item{string}{string or string vector containing keywords to search.
Note that the returned results will contain any of the keywords. It is not
possible to search for status updates that include all of the keywords.}

\item{token}{Either a temporary access token created at
\url{https://developers.facebook.com/tools/explorer} or the OAuth token
created with \code{fbOAuth}. Only tokens created for version 1.0 of the
Facebook Graph API will return results.}

\item{n}{Maximum number of posts to return.}

\item{since}{A UNIX timestamp or strtotime data value that points to
the start of the time range to be searched. For more information on the
accepted values, see: \url{http://php.net/manual/en/function.strtotime.php}}

\item{until}{A UNIX timestamp or strtotime data value that points to
the end of the time range to be searched. For more information on the
accepted values, see: \url{http://php.net/manual/en/function.strtotime.php}}

\item{api}{Version of the Graph API to query, either "1" or "2". Note that on
version 2 post search was deprecated.}
}
\description{
\code{searchFacebook} retrieves public status updates that mention a given keyword
}
\details{
The search is performed also on the text of the comments too, which explains
why some of the returned messages do not mention the string that is being
searched.

Note that only messages up to around two weeks old or less can be returned.

Public post search was deprecated with version 2.0 of the Facebook Graph API.
A temporary solution is to create a token for version 1.0. In order to do so,
go to \url{https://developers.facebook.com/tools/explorer}, select API
Version 1.0, then click on "Get Access Token". Running \code{searchFacebook}
with that token and the option api="1" will return public posts
}
\examples{
\dontrun{
## Copy and paste token created at FB Graph API Explorer
 token <- "XXXXXX"
## Searching 100 public posts that mention "facebook"
	posts <- searchFacebook( string="facebook", token=fb_oauth, n=100, api="1" )
## Searching 100 public posts that mention "facebook" from yesterday
	posts <- searchFacebook( string="facebook", token=fb_oauth, n=100,
   since = "yesterday 00:00", until = "yesterday 23:59", api="1" )
}
}
\author{
Pablo Barbera \email{pablo.barbera@nyu.edu}
}
\seealso{
\code{\link{fbOAuth}}
}

