\docType{package}
\name{Rfacebook-package}
\alias{Rfacebook}
\alias{Rfacebook-package}
\title{Access to Facebook API via R}
\description{
  This package provides a series of functions that allow R
  users to access Facebook's API to get information about
  users and posts, and collect public status updates that
  mention specific keywords.
}
\author{
  Pablo Barbera \email{pablo.barbera@nyu.edu}
}
\seealso{
  \code{\link{fbOAuth}}, \code{\link{getUsers}},
  \code{\link{getPost}}, \code{\link{searchFacebook}},
  \code{\link{updateStatus}}, \code{\link{getFriends}},
  \code{\link{getNetwork}}, \code{\link{getPage}}
}

